#var 2

+++ C++ Curse/Narrown.cpp
@@ -1,8 +1,8 @@
 #include "functions.h"
 
 int dice() {
-	const int maxValue = 6;
-	const int minValue = 1;
+	const short maxValue = 6;
+	const short minValue = 1;
 	srand(time(0));
 	int first_dice = rand() % (maxValue - minValue + 1) + minValue;
 	int second_dice = rand() % (maxValue - minValue + 1) + minValue;
